refactor: restructure files and folders to match jaen v2 structure
