+++ examples/my-gatsby-site/jaen-data/patches.txt
@@ -1,0 +1,1 @@
+https://osg.snek.at/storage/BQACAgQAAx0Ed6zoewADeGYvY7NGEG9U46vPz2iux5jZ9klUAALmGAACM9R5Ua8ba8OQycAiLwQ
