+++ examples/my-gatsby-site/package.json
