+++ examples/my-gatsby-site/src/favicon.ico
