+++ examples/my-gatsby-site/src/gatsby-plugin-jaen/components/Layout.tsx
