+++ examples/my-gatsby-site/src/gatsby-plugin-jaen/components/Logo.tsx
