+++ examples/my-gatsby-site/src/images/icon.png
