+++ examples/my-gatsby-site/src/pages/404.tsx
@@ -1,0 +1,63 @@
+import * as React from 'react'
+import {Link, HeadFC, PageProps} from 'gatsby'
+import {PageConfig} from 'jaen'
+import {LightMode, GlobalStyle} from '@chakra-ui/react'
+
+const pageStyles = {
+  color: '#232129',
+  padding: '96px',
+  fontFamily: '-apple-system, Roboto, sans-serif, serif'
+}
+const headingStyles = {
+  marginTop: 0,
+  marginBottom: 64,
+  maxWidth: 320
+}
+
+const paragraphStyles = {
+  marginBottom: 48
+}
+const codeStyles = {
+  color: '#8A6534',
+  padding: 4,
+  backgroundColor: '#FFF4DB',
+  fontSize: '1.25rem',
+  borderRadius: 4
+}
+
+const NotFoundPage: React.FC<PageProps> = () => {
+  React.useEffect(() => {
+    alert('404')
+  }, [])
+
+  return (
+    <LightMode>
+      <GlobalStyle />
+      <main style={pageStyles}>
+        <h1 style={headingStyles}>Page not found</h1>
+        <p style={paragraphStyles}>
+          Sorry 😔, we couldn’t find what you were looking for.
+          <br />
+          {process.env.NODE_ENV === 'development' ? (
+            <>
+              <br />
+              Try creating a page in <code style={codeStyles}>src/pages/</code>.
+              <br />
+            </>
+          ) : null}
+          <br />
+          <Link to="/">Go home</Link>.
+        </p>
+      </main>
+    </LightMode>
+  )
+}
+
+export default NotFoundPage
+
+export {Head} from 'jaen'
+
+export const pageConfig: PageConfig = {
+  label: 'Oops! Page not found',
+  childTemplates: ['BlogPage']
+}
