+++ examples/my-gatsby-site/src/pages/auth.tsx
@@ -1,0 +1,39 @@
+// src/App.jsx
+import React from 'react'
+import {useAuth} from 'react-oidc-context'
+
+function App() {
+  const auth = useAuth()
+
+  switch (auth.activeNavigator) {
+    case 'signinSilent':
+      return <div>Signing you in...</div>
+    case 'signoutRedirect':
+      return <div>Signing you out...</div>
+  }
+
+  if (auth.isLoading) {
+    return <div>Loading...</div>
+  }
+
+  if (auth.error) {
+    return <div>Oops... {auth.error.message}</div>
+  }
+
+  if (auth.isAuthenticated) {
+    return (
+      <div>
+        Hello {auth.user?.profile.sub}{' '}
+        <button onClick={() => void auth.removeUser()}>Log out</button>
+        <pre
+          dangerouslySetInnerHTML={{
+            __html: JSON.stringify(auth.user, null, 2)
+          }}></pre>
+      </div>
+    )
+  }
+
+  return <button onClick={() => void auth.signinRedirect()}>Log in</button>
+}
+
+export default App
