--- my-gatsby-site/.gitignore
@@ -1,4 +1,0 @@
-node_modules/
-.cache/
-public
-src/gatsby-types.d.ts
