--- my-gatsby-site/gatsby-browser.tsx
@@ -1,16 +1,0 @@
-import {GatsbyBrowser} from 'gatsby'
-import {Box, Heading} from '@chakra-ui/react'
-
-// export const wrapPageElement: GatsbyBrowser['wrapPageElement'] = ({
-//   element,
-//   props
-// }) => {
-//   return (
-//     <Box>
-//       <Box as="header" bg="gray.800" color="white" p="4">
-//         <Heading as="h1">My Site</Heading>
-//       </Box>
-//       {element}
-//     </Box>
-//   )
-// }
