--- my-gatsby-site/jaen-data/patches.txt
@@ -1,1 +1,0 @@
-https://osg.snek.at/storage/BQACAgQAAx0Ed6zoewADeGYvY7NGEG9U46vPz2iux5jZ9klUAALmGAACM9R5Ua8ba8OQycAiLwQ
